Fixed manifest package unit tests
